// src/app/core/api-v1/model/mapConfigurationRequest.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SpotDefinition } from './spotDefinition';


export interface MapConfigurationRequest { 
    mapCode?: string;
    mapName?: string;
    backgroundImageUrl?: string;
    imageWidth?: number;
    imageHeight?: number;
    spots?: Array<SpotDefinition>;
}