// src/app/core/api-v1/model/mapConfigurationResponse.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SpotDefinition } from './spotDefinition';


export interface MapConfigurationResponse { 
    mapCode?: string;
    mapName?: string;
    id?: number;
    imageWidth?: number;
    imageHeight?: number;
    backgroundImageUrl?: string;
    spots?: Array<SpotDefinition>;
}