// src/app/core/api-v1/model/productResponse.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductResponse { 
    id?: number;
    name?: string;
    type?: string;
    basePrice?: number;
}