// src/app/core/api-v1/model/spotDefinition.ts
/**
 * SpotCamp ID - Campsite Booking Platform API
 *
 * Contact: api@spotcamp.id
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Coordinate } from './coordinate';


export interface SpotDefinition { 
    id?: string;
    name?: string;
    description?: string;
    productId?: number;
    coordinates?: Array<Coordinate>;
}